Refactor secrets out of main soledad client class.
